pyproject.toml
[project]
name = "mi-proyecto"
version = "0.1.0"
description = "Promtior Challenge"
requires-python = ">=3.11"

dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "langserve>=0.2.0",
    "langchain-text-splitters>=1.1.0",
    "langchain[openai]>=1.2.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.30.0",
    "sse-starlette>=3.2.0",
    "pandas>=3.0.0",
]
